Move loading back into AddNonUMUsers leaf components; add clearGetSectionsError back to make loading reappear on tryAgain
